Partially or fully resolves a number of issues that prepare City for UI creation and runtime testing. Lots of integration code for the units I've been designing.

+++ Assets/Cities/BorderExpansionConfig.cs
@@ -32,7 +32,7 @@ namespace Assets.Cities {
         public float PreviousTileCountExponent {
             get { return _previousTileCountExponent; }
         }
-        [SerializeField] private int _previousTileCountExponent;
+        [SerializeField] private float _previousTileCountExponent;
 
         #endregion
